[MIST-564] Make EventProcessor sleep when there is no event (#564)
* [MIST-563] Make EventProcessor sleep when there is no event
This PR makes event process sleep when there is no event.

To prevent the complex synchronization of waking up threads and avoid its inefficiency, I chose "polling" mechanism. The thread will sleep for the polling interval (now its 100ms) if there is no event to process. After the polling interval, the thread wakes up and checks whether there is a data to process. If there is a data to process, then the thread starts working; if there is none, it sleeps again for the polling interval.

Adjusting polling interval will cause trade-off between overall throughput and tail latency, and further evaluation might be necessary.

Closes #563.